chore: move the copy function to util
